Changing tectonic regimes in central Costa Rica
We discuss the Cenozoic history of the Central Costa Rica forearc, between 10◦ and 9◦30′ N, based on geological
mapping, fault slip data and seismological records.The temporal variability in the regional stress-field suggests a
variable and complex deformation pattern. The first stage of deformation, from the Paleogene to early Miocene,
was controlled by an extensional regime in the forearc. Afterwards, a change to orthogonal convergence triggered
a contractional deformation and subsequent inversion of the sedimentary extensional basins in the middleupper
Miocene. Finally, a transpressive regime has been developed since the Pliocene.